Configuration has been splitted in two versions
In projects where vue is not being used, only `nosolosoftware` has to be
imported. It allows us to not depend on eslint-config-vue in projects
where we are using this styleguide but vue is not present. For those
cases, where vuejs is being used `nosolosoftware/vue` has also to be
imported